YT-21148 Cosmetics: Do not expose `TSerializableColumnSchema` and improve certain error messages commit_hash:74e4f9bbad5e6c60e8b3b2f6b1c924be165d59ce
